Added support for Awkward Arrays to support obs from JSON like sources
Alleviates issues of having zero as an arc number as now the preferred input method is to use awkward arrays rather than sparse matrices. Note that the rationale for awkward arrays is the same as using list of lists rather than spase, since nonzero in obs are all left contiguous. Unclear whether awkward arrays are faster than list of lists, but they do support array indexing which is handy.